Introduced an appropriate scaling factor and applied the appropriate mathematics for the offset of the origin between Box2D and OpenGL. Therefore the transform component now accurately and appropriately can be deemed to use pixels, of which there is a current scaling factor of 100 pixels per metre.

+++ Engine/src/elements/RenderElement.cpp
@@ -12,7 +12,7 @@ namespace CoreRenderElement
 	{
 	}
 
-	bool RenderElement::initialize(const std::string& windowName, int width, int height, SDL_WindowFlags window_flags, int vsync)
+	bool RenderElement::initialize(const std::string& windowName, float width, float height, SDL_WindowFlags window_flags, int vsync)
 	{
 		auto corePtr = m_core.lock();
 		if (corePtr)
